+++ code/modules/reagents/reagent_containers/glass.dm
@@ -33,6 +33,13 @@
 /obj/item/weapon/reagent_containers/glass/afterattack(obj/target, mob/user, proximity)
 	if((!proximity) || !check_allowed_items(target)) return
 
+	if(istype(target, /obj/machinery/brewery)) // to fix bag when splashing fresh wine imedeatly
+		if (!target.is_open_container())
+			return
+
+	if(istype(target, /obj/effect/spacevine) && target.reagents)	//harwesting reagents from spacevine
+		return
+
 	if(ismob(target) && target.reagents && reagents.total_volume)
 		var/mob/M = target
 		var/R
@@ -235,9 +242,12 @@
 	possible_transfer_amounts = list(10,20,30,50,70)
 	volume = 70
 	flags = OPENCONTAINER
+	var/cyborg = 0
 
 /obj/item/weapon/reagent_containers/glass/bucket/attackby(var/obj/D, mob/user as mob, params)
 	if(isprox(D))
+		if(cyborg)
+			return
 		user << "<span class='notice'>You add [D] to [src].</span>"
 		qdel(D)
 		user.put_in_hands(new /obj/item/weapon/bucket_sensor)
